+++ src/family_tree_app/model/family_tree/FamilyTree.java
@@ -1,0 +1,6 @@
+package family_tree_app.model.family_tree;
+
+import java.io.Serializable;
+
+public interface FamilyTree<F> extends Serializable {
+}
